feat: address comments
