+++ test/standardizedToken.js
@@ -4,11 +4,10 @@ const chai = require('chai');
 const { ethers } = require('hardhat');
 const {
     Contract,
-    utils: { splitSignature, keccak256, toUtf8Bytes },
-    constants: { MaxUint256, AddressZero },
+    utils: { keccak256, toUtf8Bytes },
 } = ethers;
 const { expect } = chai;
-const { getRandomBytes32, getChainId, expectRevert, getGasOptions } = require('../scripts/utils');
+const { getRandomBytes32, expectRevert, getGasOptions } = require('./utils');
 const { deployContract } = require('../scripts/deploy');
 
 const StandardizedToken = require('../artifacts/contracts/token-implementations/StandardizedToken.sol/StandardizedToken.json');
@@ -25,15 +24,12 @@ describe('StandardizedToken', () => {
     let token;
     let tokenProxy;
 
-    let owner, user;
+    let owner;
 
     before(async () => {
         const wallets = await ethers.getSigners();
         owner = wallets[0];
-        user = wallets[1];
-    });
 
-    beforeEach(async () => {
         standardizedToken = await deployContract(owner, 'StandardizedToken');
         standardizedTokenDeployer = await deployContract(owner, 'StandardizedTokenDeployer', [standardizedToken.address]);
 
@@ -44,16 +40,9 @@ describe('StandardizedToken', () => {
         token = new Contract(tokenAddress, StandardizedToken.abi, owner);
         tokenProxy = new Contract(tokenAddress, StandardizedTokenProxy.abi, owner);
 
-        await standardizedTokenDeployer.deployStandardizedToken(
-            salt,
-            owner.address,
-            owner.address,
-            name,
-            symbol,
-            decimals,
-            mintAmount,
-            owner.address,
-        );
+        await standardizedTokenDeployer
+            .deployStandardizedToken(salt, owner.address, owner.address, name, symbol, decimals, mintAmount, owner.address)
+            .then((tx) => tx.wait());
     });
 
     describe('Standardized Token Proxy', () => {
@@ -92,213 +81,6 @@ describe('StandardizedToken', () => {
 
             const params = '0x';
             await expectRevert((gasOptions) => implementation.setup(params, gasOptions), token, 'NotProxy');
-        });
-    });
-
-    describe('ERC20 Basics', () => {
-        it('should increase and decrease allowance', async () => {
-            const initialAllowance = await token.allowance(user.address, owner.address);
-            expect(initialAllowance).to.eq(0);
-
-            await expect(token.connect(user).increaseAllowance(owner.address, MaxUint256))
-                .to.emit(token, 'Approval')
-                .withArgs(user.address, owner.address, MaxUint256);
-
-            const increasedAllowance = await token.allowance(user.address, owner.address);
-            expect(increasedAllowance).to.eq(MaxUint256);
-
-            await expect(token.connect(user).decreaseAllowance(owner.address, MaxUint256))
-                .to.emit(token, 'Approval')
-                .withArgs(user.address, owner.address, 0);
-
-            const finalAllowance = await token.allowance(user.address, owner.address);
-            expect(finalAllowance).to.eq(0);
-        });
-
-        it('should revert on approve with invalid owner or sender', async () => {
-            await expectRevert(
-                (gasOptions) => token.connect(owner).transferFrom(AddressZero, owner.address, 0, gasOptions),
-                token,
-                'InvalidAccount',
-            );
-
-            await expectRevert(
-                (gasOptions) => token.connect(user).increaseAllowance(AddressZero, MaxUint256, gasOptions),
-                token,
-                'InvalidAccount',
-            );
-        });
-
-        it('should revert on transfer to invalid address', async () => {
-            const initialAllowance = await token.allowance(user.address, owner.address);
-            expect(initialAllowance).to.eq(0);
-
-            await expect(token.connect(user).increaseAllowance(owner.address, MaxUint256))
-                .to.emit(token, 'Approval')
-                .withArgs(user.address, owner.address, MaxUint256);
-
-            const increasedAllowance = await token.allowance(user.address, owner.address);
-            expect(increasedAllowance).to.eq(MaxUint256);
-
-            const amount = 100;
-
-            await expectRevert(
-                (gasOptions) => token.connect(owner).transferFrom(user.address, AddressZero, amount, gasOptions),
-                token,
-                'InvalidAccount',
-            );
-        });
-
-        it('should revert mint or burn to invalid address', async () => {
-            const amount = 100;
-            await expectRevert((gasOptions) => token.mint(AddressZero, amount, gasOptions), token, 'InvalidAccount');
-            await expectRevert((gasOptions) => token.burn(AddressZero, amount, gasOptions), token, 'InvalidAccount');
-        });
-    });
-
-    describe('ERC20 Permit', () => {
-        it('should set allowance by verifying permit', async () => {
-            const deadline = Math.floor(Date.now() / 1000) + 1000;
-            const allowance = 10000;
-
-            const nonce = await token.nonces(owner.address);
-
-            const signature = splitSignature(
-                await owner._signTypedData(
-                    {
-                        name,
-                        version: '1',
-                        chainId: await getChainId(),
-                        verifyingContract: token.address,
-                    },
-                    {
-                        Permit: [
-                            { name: 'owner', type: 'address' },
-                            { name: 'spender', type: 'address' },
-                            { name: 'value', type: 'uint256' },
-                            { name: 'nonce', type: 'uint256' },
-                            { name: 'deadline', type: 'uint256' },
-                        ],
-                    },
-                    {
-                        owner: owner.address,
-                        spender: user.address,
-                        value: allowance,
-                        nonce,
-                        deadline,
-                    },
-                ),
-            );
-
-            await expect(
-                token
-                    .connect(owner)
-                    .permit(owner.address, user.address, allowance, deadline, signature.v, signature.r, signature.s, { gasLimit: 8000000 }),
-            )
-                .to.emit(token, 'Approval')
-                .withArgs(owner.address, user.address, allowance);
-
-            expect(await token.nonces(owner.address)).to.equal(nonce.add(1));
-
-            expect(await token.allowance(owner.address, user.address)).to.equal(allowance);
-        });
-
-        it('should revert if permit is expired', async () => {
-            const deadline = 100;
-            const allowance = 10000;
-
-            const signature = splitSignature(
-                await user._signTypedData(
-                    {
-                        name: 'test',
-                        version: '1',
-                        chainId: await getChainId(),
-                        verifyingContract: token.address,
-                    },
-                    {
-                        Permit: [
-                            { name: 'owner', type: 'address' },
-                            { name: 'spender', type: 'address' },
-                            { name: 'value', type: 'uint256' },
-                            { name: 'nonce', type: 'uint256' },
-                            { name: 'deadline', type: 'uint256' },
-                        ],
-                    },
-                    {
-                        owner: user.address,
-                        spender: owner.address,
-                        value: allowance,
-                        nonce: 0,
-                        deadline,
-                    },
-                ),
-            );
-
-            await expectRevert(
-                (gasOptions) =>
-                    token
-                        .connect(owner)
-                        .permit(user.address, owner.address, allowance, deadline, signature.v, signature.r, signature.s, gasOptions),
-                token,
-                'PermitExpired',
-            );
-        });
-
-        it('should revert if signature is incorrect', async () => {
-            const deadline = (1000 + Date.now() / 1000) | 0;
-            const allowance = 10000;
-
-            const signature = splitSignature(
-                await user._signTypedData(
-                    {
-                        name: 'test',
-                        version: '1',
-                        chainId: await getChainId(),
-                        verifyingContract: token.address,
-                    },
-                    {
-                        Permit: [
-                            { name: 'owner', type: 'address' },
-                            { name: 'spender', type: 'address' },
-                            { name: 'value', type: 'uint256' },
-                            { name: 'nonce', type: 'uint256' },
-                            { name: 'deadline', type: 'uint256' },
-                        ],
-                    },
-                    {
-                        owner: user.address,
-                        spender: owner.address,
-                        value: allowance,
-                        nonce: 0,
-                        deadline,
-                    },
-                ),
-            );
-
-            await expectRevert(
-                (gasOptions) =>
-                    token
-                        .connect(owner)
-                        .permit(user.address, owner.address, allowance, deadline, signature.v, signature.r, MaxUint256, gasOptions),
-                token,
-                'InvalidS',
-            );
-
-            await expectRevert(
-                (gasOptions) =>
-                    token.connect(owner).permit(user.address, owner.address, allowance, deadline, 0, signature.r, signature.s, gasOptions),
-                token,
-                'InvalidV',
-            );
-
-            await expectRevert(
-                (gasOptions) =>
-                    token
-                        .connect(owner)
-                        .permit(owner.address, owner.address, allowance, deadline, signature.v, signature.r, signature.s, gasOptions),
-                token,
-                'InvalidSignature',
-            );
         });
     });
 });
